requirements.txt
pandas
azure-cognitiveservices-speech
python-dotenv
git+https://github.com/weatherapicom/python.git#egg=weatherapipython
requests
transformers
torch
tiktoken
protobuf
sentencepiece
dateparser
google-genai
pytest
fastapi
jinja2
uvicorn
fastapi[standard]
pydub
sqlalchemy
sqlalchemy.orm